Make sure any installation or runtime errors that might be caused by the particle-usb library would only affect mesh and usb commands
